move executino upsert tests
